fix: move logic into sponsorship related with flatRate
