+++ apps/trips/src/common/application/application.module.ts
@@ -5,12 +5,10 @@ import { ApplicationRepository } from './application.repository';
 import { ApplicationResolver } from './application.resolver';
 import { ApplicationService } from './application.service';
 import { TripModule } from '../trip/trip.module';
-import { ConfigurationModule } from '../configuration/configuration.module';
 
 @Module({
   imports: [
     TripModule,
-    ConfigurationModule,
     MongooseModule.forFeature([
       {
         name: Application.name,
