+++ apps/trips/src/common/application/application.resolver.ts
@@ -1,11 +1,4 @@
-import {
-  Resolver,
-  Query,
-  Args,
-  Mutation,
-  ResolveField,
-  Float
-} from '@nestjs/graphql';
+import { Resolver, Query, Args, Mutation } from '@nestjs/graphql';
 import { Application } from './graphql/types/application.type';
 import { FilterInput } from '@shared/graphql/inputs/graphql-filter.input';
 import { graphQlIdArgOption } from '@shared/graphql/types/graphql-delete-mutation-options.type';
@@ -185,11 +178,5 @@ export class ApplicationResolver {
         reasonRejected: rejectApplicationInput.reasonRejected
       }
     });
-  }
-
-  // Field resolvers
-  @ResolveField(() => Float)
-  public async flatRate() {
-    return this.service.flatRate();
   }
 }
